server/requirements.txt
fastapi==0.109.0
uvicorn==0.26.0
websockets==12.0
Python-IO==0.3
google-api-python-client==2.114.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-auth==2.26.2
google-cloud==0.34.0